Moved GCAllocator to its own file & added #includes
allocator (the class) now resides in gcallocator.h & gcallocator.cpp

#includes were added to GCStatistics and allocator

Copyright notice was copy pasted.
